test drive rentals json data parsing
I'm totally seeing a scala influence here with building up Structs for
all the different things, as opposed to letting them remain as simple
hash/array data structures.

In scala, I use case classes extensively for these simple data objects.
They typically have no behavior. Any behavior they do have is usually
along the lines of convenience methods around data transformation and
access.

There's also a bit of clojure influence in that the data transformations
are confined to one place. At the place where the data is parsed, we
ensure we're sanitized, so that downstream you can count on a Rental or
Season being correct. The ParsesData class is like a sanitize function
I would write in clojure, and using it as a nested class means I can test
it independently from anything else.

The actual parsing work is big and messy, but pretty easy to follow and
a good candidate for refactoring. While refactoring, I'll likely add
fuzz testing to the #from_stringified_hash method to make it more robust.

For now though, I have a faked integration test that needs to pass for
real before we can make the code pretty.