refactor sales tax
Remove it from the Rentals class's state. It now gets
injected into the estimate method where it's needed.

This also means we're not keeping a sales tax constant
hanging around -- we default to zero, which is basically
a tax-free estimate.

It also makes the tax adjustment clearer in the 
acceptance test.